+++ tests/test_google_photos.py
@@ -12,7 +12,7 @@ from typing import Any, Tuple, Dict, Union
 from unittest import mock
 
 from app.google_photos import GooglePhotos, GoogleApiResponseNG
-from tests.lib.ulit import load_json, delete_env
+from tests.lib.utils import load_json, delete_env
 
 JSON_DIR = f'{os.path.dirname(__file__)}/json'
 STATIC_CONTENT_DIR = f'{os.path.dirname(__file__)}/static_content'
